requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
langgraph>=0.0.10
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.8
chromadb>=0.4.22
pydantic>=2.6.0
httpx>=0.27.0
pytest>=8.0.0
python-dotenv>=1.0.1
aiofiles>=23.2.1
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
networkx>=3.0
apscheduler>=3.10.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
